software/scanner/requirements.txt
python-nmap>=0.6.1
luma.oled>=3.2.1
configparser>=3.5.2
coloredlogs>=10.0
netaddr>=0.7.19
spidev>=3.4.0
json2html>=1.3.0
dirtyjson>=1.0.7
tablib>=1.1.0
flask-login>=0.5.0
flask-sqlalchemy>=0.8.0
Flask-SSLify>=0.1.5


